Document behaviour of skips in beamercolorbox (fixes issue #1)
While this is only a documentation fix, we have little other option.
Till exposed the raw TeX lengths in the interface he chose, and to
alter that now would break a lot of existing code. So the best we
can do is note that there is an interaction: if you use "leftskip"
or "rightskip" then you are on your own!