Enables customization of the note page look.
The colors that were used in the note page were half hard coded and half inherited from other values. This has the unintended consequence to give some conflicting values. For example, setting a normal text foreground to white meant that the note page foreground was white with no way of setting a note page background to anything but white. The only recourse was to change the note page foreground to display nicely on white.

Customization for the note page title, date, and background is added.

The mini slide also inherits the foreground and background color of normal text.