Fix incorrect Beamer template names for 'structureenv' (#862)
As per Section 12.2 of the Beamer User Guide for version 3.70, the
Beamer template names for 'structureenv' should be 'structure begin' and
'structure end'.  However, some of the Beamer package's source files are
not using these exact names.

An issue caused by this is that, in article mode, text in 'structureenv'
is not typeset in boldface as specified also by Section 12.2.  A minimal
working example is:

    \documentclass{article}
    \usepackage{beamerarticle}
    \begin{document}
    \begin{frame}
        \structure{This text should be typeset in boldface.}
    \end{frame}
    \end{document}

Signed-off-by: Yuan Liao <liaoyuan@gmail.com>